slurm-conductor/requirements.txt
ops
psycopg2-binary
